+++ src/index/entry.rs
@@ -24,13 +24,13 @@ impl Entry for BlockHash {
 
 pub(crate) struct InscriptionEntry {
   pub(crate) fee: u64,
-  pub(crate) height: u64,
-  pub(crate) number: u64,
+  pub(crate) height: u32,
+  pub(crate) number: u32,
   pub(crate) sat: Option<Sat>,
   pub(crate) timestamp: u32,
 }
 
-pub(crate) type InscriptionEntryValue = (u64, u64, u64, u128, u32);
+pub(crate) type InscriptionEntryValue = (u64, u32, u32, u128, u32);
 
 impl Entry for InscriptionEntry {
   type Value = InscriptionEntryValue;
